MonitorService starts the monitor
